+++ studio/src/app/utils/editor/parse-slides.utils.tsx
@@ -13,16 +13,8 @@ export class ParseSlidesUtils {
                 return;
             }
 
-            if (SlideTemplate[slide.data.template.toUpperCase()] === SlideTemplate.TITLE) {
-                resolve(await this.parseSlideElement(slide, 'deckgo-slide-title'));
-            } else if (SlideTemplate[slide.data.template.toUpperCase()] === SlideTemplate.CONTENT) {
-                resolve(await this.parseSlideElement(slide, 'deckgo-slide-content'));
-            } else if (SlideTemplate[slide.data.template.toUpperCase()] === SlideTemplate.SPLIT) {
-                resolve(await this.parseSlideElement(slide, 'deckgo-slide-split'));
-            } else if (SlideTemplate[slide.data.template.toUpperCase()] === SlideTemplate.GIF) {
-                resolve(await this.parseSlideElement(slide, 'deckgo-slide-gif'));
-            } else if (SlideTemplate[slide.data.template.toUpperCase()] === SlideTemplate.AUTHOR) {
-                resolve(await this.parseSlideElement(slide, 'deckgo-slide-author'));
+            if (SlideTemplate[slide.data.template.toUpperCase()]) {
+                resolve(await this.parseSlideElement(slide, `deckgo-slide-${SlideTemplate[slide.data.template.toUpperCase()].toLowerCase()}`));
             } else {
                 resolve(null);
             }
@@ -54,7 +46,7 @@ export class ParseSlidesUtils {
             const SlideElement: string = slideTag;
 
             // @ts-ignore
-            const result: any = <SlideElement slide_id={slide.id} style={style} src={src} custom-background={customBackground} img-src={imgSrc} img-alt={imgAlt}>
+            const result: any = <SlideElement key={slide.id} slide_id={slide.id} style={style} src={src} custom-background={customBackground} img-src={imgSrc} img-alt={imgAlt}>
                 {content}
             </SlideElement>;